+++ Integration/dbo/Views/vyuCPCurrentCashBids.sql
@@ -1,5 +1,10 @@
-﻿IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vyuCPCurrentCashBids')
+﻿IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vwCPCurrentCashBids')
+	DROP VIEW vwCPCurrentCashBids
+GO
+
+IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vyuCPCurrentCashBids')
 	DROP VIEW vyuCPCurrentCashBids
+GO
 
 -- GRAINS DEPENDENT
 IF  (SELECT TOP 1 ysnUsed FROM ##tblOriginMod WHERE strPrefix = 'GR' and strDBName = db_name()	) = 1 and
