+++ Integration/dbo/Views/vyuCPGABusinessSummary.sql
@@ -1,4 +1,7 @@
-﻿IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vyuCPGABusinessSummary')
+﻿IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vwCPGABusinessSummary')
+	DROP VIEW vwCPGABusinessSummary
+GO
+IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vyuCPGABusinessSummary')
 	DROP VIEW vyuCPGABusinessSummary
 GO
 -- GRAINS DEPENDENT
