+++ Integration/dbo/Views/vyuCPPurchases.sql
@@ -1,4 +1,7 @@
-﻿IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vyuCPPurchases')
+﻿IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vwCPPurchases')
+	DROP VIEW vwCPPurchases
+GO
+IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vyuCPPurchases')
 	DROP VIEW vyuCPPurchases
 GO
 -- AG VIEW
