+++ Integration/dbo/Views/vyuCPSettlements.sql
@@ -1,4 +1,7 @@
-﻿IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vyuCPSettlements')
+﻿IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vwCPSettlements')
+	DROP VIEW vwCPSettlements
+GO
+IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vyuCPSettlements')
 	DROP VIEW vyuCPSettlements
 GO
 -- GRAINS DEPENDENT
